\name{bay.cdf}
\alias{bay.cdf}
\title{Bay cumulative density function, corresponding to -B(t+)/B(t-), where B(t+) (resp. B(t-)) is the maximum (resp.minimum) of B(t)/(t*(1-t)).}
\usage{
  bay.cdf(x)
}
\arguments{
  \item{x}{value to test}
}
\description{
  Bay cumulative density function, corresponding to
  -B(t+)/B(t-), where B(t+) (resp. B(t-)) is the maximum
  (resp.minimum) of B(t)/(t*(1-t)).
}
\author{
  X. Bay
}

