\name{null.param.cdf}
\alias{null.param.cdf}
\title{Build the null CDF (cumulative density function) for a given statistic, for arbitrary length and autocorrelation sequence.}
\usage{
  null.param.cdf(stat, N, rho)
}
\arguments{
  \item{stat}{statistic used}

  \item{N}{length of the target sequence to be tested}

  \item{rho}{autocorrelation (1st coeff) of the target
  sequence to test}
}
\description{
  Build the null CDF (cumulative density function) for a
  given statistic, for arbitrary length and autocorrelation
  sequence.
}

